% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Document package.}
\usage{
document(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
This function calls \code{\link[devtools:document]{devtools::document()}} to update NAMESPACE and documentation
and then calls \code{\link[=write_dependencies]{write_dependencies()}} to update the DESCRIPTION file.
}
\examples{
\dontrun{
document()
}
}
