% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Returns all packages used by the package.}
\usage{
get_dependencies(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\value{
A character vector with the packages used in the package.
}
\description{
Scans all files in R/ for used packages (searching for :: syntax) and scans
NAMESPACE file for import and importFrom.
}
\details{
Make sure to import all dependencies or use the explicit \code{::} syntax when
calling functions from other packages and to generate the \code{NAMESPACE} first.
}
\examples{
\dontrun{
get_dependencies
}
}
