% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dependencies.r
\name{update_dependencies}
\alias{update_dependencies}
\title{Update Imports section in DESCRIPTION file.}
\usage{
update_dependencies()
}
\value{
Nothing.
}
\description{
Scans all files in R/ for used packages (searching for :: syntax) and scans
NAMESPACE file for import and importFrom. Then writes dependencies to Imports
section in DESCRIPTION file.
}
\examples{
\dontrun{
update_dependencies()
}
}
