% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dependencies.r
\name{write_dependencies}
\alias{write_dependencies}
\title{Update Imports section in DESCRIPTION file.}
\usage{
write_dependencies(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Scans all files in R/ for used packages (searching for :: syntax) and scans
NAMESPACE file for import and importFrom. Then writes dependencies to Imports
section in DESCRIPTION file.
}
\details{
Make sure to import all dependencies or use the explicit \code{::} syntax when
calling functions from other packages and to generate the \code{NAMESPACE} first.
}
\examples{
\dontrun{
write_dependencies()
}
}
