% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Document package}
\usage{
document(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name.  See
\code{\link{as.package}} for more information}
}
\description{
Thsi function calls \code{\link[devtools:document]{devtools::document()}} to update NAMESPACE and documentation
and then calls \code{\link[=write_dependencies]{write_dependencies()}} to update the DESCRIPTION file.
}
\examples{
\dontrun{
document()
}
}
