% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Return all dependency packages.}
\usage{
get_dependencies(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\value{
A character vector with the packages used in the package.
}
\description{
Scans all files in \code{R/} for used packages (searching for \code{::} syntax) and scans
\code{NAMESPACE} file for import and importFrom statements.
}
\details{
Make sure to import all dependencies or use the explicit \code{::} syntax when
calling functions from other packages and to generate the \code{NAMESPACE} first.
}
\examples{
\dontrun{
get_dependencies()
}
}
