% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dependencies.r
\name{write_dependencies}
\alias{write_dependencies}
\title{Write dependency packages to DESCRIPTION.}
\usage{
write_dependencies(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[=as.package]{as.package()}} for more information.}
}
\description{
Scans all files in \code{R/} for used packages (searching for \code{::} syntax) and scans
\code{NAMESPACE} file for import and importFrom statements.
Then writes dependencies to \code{Imports} section in \code{DESCRIPTION} file.
}
\details{
Make sure to import all dependencies or use the explicit \code{::} syntax when
calling functions from other packages and to generate the \code{NAMESPACE} first.
}
\examples{
\dontrun{
write_dependencies()
}
}
